Issue #00 chore: setup devops code script updated

+++ pipelines/setupDevopsCode.sh
@@ -4,7 +4,10 @@ set -ex
 
 commit_ref=${1:-release-1.7}
 sunbird_devops_dir_name=sunbird-devops
-[ -d $sunbird_devops_dir_name ] || git clone https://github.com/project-sunbird/sunbird-devops -b $commit_ref; exit 0;
+if [ -d $sunbird_devops_dir_name ];then
+    git clone https://github.com/project-sunbird/sunbird-devops -b $commit_ref
+    exit 0
+fi
 cd $sunbird_devops_dir_name
 git reset --hard
 git clean -fd
